requirements.txt
aiosmtplib==1.1.7
anyio==3.7.1
asgiref==3.8.1
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.8
Django==5.0.6
django-allauth==0.63.3
django-appconf==1.0.6
django-autocomplete-light==3.11.0
django-select2==8.2.0
djangorestframework==3.15.2
h11==0.12.0
httpcore==0.13.7
httpx==0.19.0
idna==3.7
jwt==1.3.1
messages==0.8.0
mysqlclient==2.2.4
pillow==10.3.0
pycparser==2.22
PyJWT==2.8.0
requests==2.32.3
rfc3986==1.5.0
sniffio==1.3.1
sqlparse==0.5.0
urllib3==2.2.2
validus==0.3.0
